+++ src/gs1.js
@@ -6,10 +6,14 @@ export {
   calculateSsccCheckDigit,
   createGs1DigitalLink,
   createGs1ElementString,
+  getGs1AiInfo,
+  getSupportedGs1Ais,
   GS1_FNC1_SEPARATOR,
   parseGs1DigitalLink,
   parseGs1HumanReadable,
   validateGs1CheckDigit,
+  validateGs1Elements,
+  validateGs1ElementString,
   validateGtinCheckDigit,
   validateSsccCheckDigit
 } from "./gs1/index.js";